handle a few failure states

+++ server/tmserver/world.py
@@ -472,9 +472,7 @@ class GameWorld:
             'target_room_name': target_room.shortname})
 
         with get_db().atomic():
-            exits = current_room.get_data('exits')
-            if exits is None:
-                exits = {}
+            exits = current_room.get_data('exits', {})
             exits[direction] = here_exit.shortname
             current_room.set_data('exits', exits)
             cls.put_into(current_room, here_exit)
@@ -490,9 +488,7 @@ class GameWorld:
                 'target_room_name': current_room.shortname})
             rev_dir = REVERSE_DIRS[direction]
             with get_db().atomic():
-                exits = target_room.get_data('exits')
-                if exits is None:
-                    exits = {}
+                exits = target_room.get_data('exits', {})
                 exits[rev_dir] = there_exit.shortname
                 target_room.set_data('exits', exits)
                 cls.put_into(target_room, there_exit)
@@ -580,8 +576,14 @@ class GameWorld:
         # shortname of a room in the database. In the future we might need
         # fuzzy matching but for now I think moves are largely programmatic?
         room = GameObject.get_or_none(GameObject.shortname==action_args)
-        cls.put_into(room, sender_obj)
-        cls.user_hears(sender_obj, sender_obj, 'You materialize in a new place!')
+        if room:
+            if sender_obj == room:
+                cls.user_hears(sender_obj, sender_obj, "You can't move to yourself.")
+            else: 
+                cls.put_into(room, sender_obj)
+                cls.user_hears(sender_obj, sender_obj, 'You materialize in a new place!')
+        else:
+            cls.user_hears(sender_obj, sender_obj, "Can't figure out what you meant to move to.")
 
     @classmethod
     def handle_go(cls, sender_obj, action_args):
@@ -605,7 +607,7 @@ class GameWorld:
 
         direction = action_args
         current_room = sender_obj.contained_by
-        exits = current_room.get_data('exits')
+        exits = current_room.get_data('exits', {})
         if direction not in exits:
             cls.user_hears(sender_obj, sender_obj, 'You cannot go that way.')
             return
@@ -656,6 +658,8 @@ class GameWorld:
 
     @classmethod
     def put_into(cls, outer_obj, inner_obj):
+        if outer_obj == inner_obj:
+            raise ClientException('Cannot put something into itself.')
         if inner_obj.contained_by:
             old_outer_obj = inner_obj.contained_by
             Contains.delete().where(Contains.inner_obj==inner_obj).execute()
